Remove duplicate `UpsertSubscriptionsDto` entities and builders
