+++ src/domain/interfaces/notifications.datasource.interface.ts
@@ -1,4 +1,4 @@
-import type { UpsertSubscriptionsDto } from '@/routes/notifications/v1/entities/upsert-subscriptions.dto.entity';
+import type { UpsertSubscriptionsDto } from '@/domain/notifications/v2/entities/upsert-subscriptions.dto.entity';
 import type { NotificationType } from '@/domain/notifications/v2/entities/notification-type.entity';
 import type { UUID } from 'crypto';
